docs:add new files and modify some files(#39519)
