+++ extra/patterns/index.js
@@ -1,0 +1,4 @@
+import diamond from "./diamond";
+import triangle from "./eql-triangle";
+
+export { diamond, triangle };
