+++ extra/utils/permissions.js
@@ -1,24 +1,5 @@
-const permissions = {
-  getUsers: {
-    all: ["head-trainer"],
-    read: ["trainee", "trainer"],
-    write: ["trainer"],
-    delete: []
-  },
-  getUsers1: {
-    all: ["head-trainer", "trainee"],
-    read: ["trainee", "trainer"],
-    write: ["trainer"],
-    delete: []
-  },
-  getUsers2: {
-    all: ["head-trainer", "trainee", "trainer"],
-    read: ["trainee", "trainer"],
-    write: ["trainer"],
-    delete: []
-  }
-};
-function hasPermission(moduleName, role, permissionType) {
+import { permissions } from "/home/trainode26/Projects/express-server/constants";
+export default function hasPermission(moduleName, role, permissionType) {
   if (permissions.hasOwnProperty(moduleName)) {
     if (permissions[moduleName][permissionType].includes(role)) {
       console.log(true);
@@ -29,4 +10,3 @@ function hasPermission(moduleName, role, permissionType) {
     console.log(false);
   }
 }
-hasPermission("getUsers1", "head-trainer", "all");
